refactor: split AlpmManager into partial classes for better maintainability
